Split test file according to using html file
